move event listeners to render functions
