+++ src/app/interfaces/IRoute.ts
@@ -1,7 +1,6 @@
 interface IRoute {
   path: string;
   renderFunction(el: HTMLElement): void;
-  script?(): void;
 }
 
 export default IRoute;
